+++ tests/test_device_find.py
@@ -179,7 +179,7 @@ def test_device_query(user: UserClient):
     snapshot, _ = user.post(conftest.file('basic.snapshot'), res=Snapshot)
     i, _ = user.get(res=Device)
     assert i['url'] == '/devices/'
-    assert i['items'][0]['url'] == '/devices/{}'.format(snapshot['device']['id'])
+    assert i['items'][0]['url'] == '/devices/%s' % snapshot['device']['devicehubID']
     pc = next(d for d in i['items'] if d['type'] == 'Desktop')
     assert len(pc['actions']) == 4
     assert len(pc['components']) == 3