+++ tests/test_lot.py
@@ -365,13 +365,14 @@ def test_lot_post_add_remove_device_view(app: Devicehub, user: UserClient):
         db.session.add(device)
         db.session.commit()
         device_id = device.id
+        devicehub_id = device.devicehub_id
     parent, _ = user.post(({'name': 'lot'}), res=Lot)
     lot, _ = user.post({},
                        res=Lot,
                        item='{}/devices'.format(parent['id']),
                        query=[('id', device_id)])
     assert lot['devices'][0]['id'] == device_id, 'Lot contains device'
-    device, _ = user.get(res=Device, item=device_id)
+    device, _ = user.get(res=Device, item=devicehub_id)
     assert len(device['lots']) == 1
     assert device['lots'][0]['id'] == lot['id'], 'Device is inside lot'